+++ src/components/global/header/headerPanel.ts
@@ -1,5 +1,5 @@
 import { LitElement, html } from 'lit';
-import { customElement, property, state } from 'lit/decorators.js';
+import { customElement, property } from 'lit/decorators.js';
 import { classMap } from 'lit/directives/class-map.js';
 import HeaderPanelScss from './headerPanel.scss';
 import '@kyndryl-design-system/shidoka-foundation/components/icon';
@@ -7,7 +7,7 @@ import '@kyndryl-design-system/shidoka-foundation/components/icon';
 import closeIcon from '@carbon/icons/es/close/24';
 
 /**
- * Header fly-out panel.
+ * DEPRECATED. Header fly-out panel.
  * @slot unnamed - Slot for panel content.
  * @slot button - Slot for button icon.
  */
@@ -35,12 +35,6 @@ export class HeaderPanel extends LitElement {
   @property({ type: String })
   closeText = 'Close Panel';
 
-  /**
-   * A generated unique id
-   * @ignore
-   */
-  @state() private _id = crypto.randomUUID();
-
   override render() {
     const buttonClasses = {
       interactive: true,
@@ -59,20 +53,20 @@ export class HeaderPanel extends LitElement {
         title=${this.openText}
         aria-label=${this.openText}
         @click=${this.togglePanel}
-        aria-controls=${this._id}
+        aria-controls="panel"
         aria-expanded=${this.open}
       >
         <slot name="button"></slot>
       </button>
 
-      <div id=${this._id} class="${classMap(panelClasses)}" tabindex="-1">
+      <div id="panel" class="${classMap(panelClasses)}" tabindex="-1">
         <div class="heading">
           <button
             class="${classMap(buttonClasses)}"
             title=${this.closeText}
             aria-label=${this.closeText}
             @click=${this.togglePanel}
-            aria-controls=${this._id}
+            aria-controls="panel"
             aria-expanded=${this.open}
           >
             <kd-icon .icon=${closeIcon}></kd-icon>