+++ src/components/reusable/daterangepicker/daterangepicker.ts
@@ -11,7 +11,6 @@ import DateRangePickerScss from './daterangepicker.scss';
  * @prop {string} minDate - Mimimum date in YYYY-MM-DD format. If the value isn't a possible date string in the format, then the element has no minimum date value.
  * @prop {string} maxDate - Maximum date in YYYY-MM-DD format. If the value isn't a possible date string in the format, then the element has no maximum date value.
  * @slot unnamed - Slot for label text.
- * @csspart label - Styles the label.
  */
 @customElement('kyn-date-range-picker')
 export class DateRangePicker extends LitElement {
@@ -120,7 +119,6 @@ export class DateRangePicker extends LitElement {
   override render() {
     return html`
       <label
-        part="label"
         class="datepicker-label-text"
         for=${this.name}
         ?disabled=${this.disabled}